+++ src/lsp.ts
@@ -192,18 +192,19 @@ function activate(context: vscode.ExtensionContext): LanguageClient {
     })
     context.subscriptions.push(client.start());
 
-    const jarEventEmitter: vscode.EventEmitter<vscode.Uri> = new vscode.EventEmitter();
-    const contentsRequest = new RequestType<string, string, string, vscode.CancellationToken>('clojure/dependencyContents');
-    const textDocumentContentProvider: vscode.TextDocumentContentProvider = {
-        onDidChange: jarEventEmitter.event,
-        provideTextDocumentContent: async (uri: vscode.Uri, token: vscode.CancellationToken): Promise<string> => {
-            const v = await client.sendRequest<any, string, string, vscode.CancellationToken>(contentsRequest,
-                { uri: decodeURIComponent(uri.toString()) },
-                token);
-            return v || '';
-        }
-    };
-    context.subscriptions.push(vscode.workspace.registerTextDocumentContentProvider('jar', textDocumentContentProvider));
+    // TODO: We don't need two jar content providers, when would this one be needed?
+    //const jarEventEmitter: vscode.EventEmitter<vscode.Uri> = new vscode.EventEmitter();
+    //const contentsRequest = new RequestType<string, string, string, vscode.CancellationToken>('clojure/dependencyContents');
+    //const textDocumentContentProvider: vscode.TextDocumentContentProvider = {
+    //    onDidChange: jarEventEmitter.event,
+    //    provideTextDocumentContent: async (uri: vscode.Uri, token: vscode.CancellationToken): Promise<string> => {
+    //        const v = await client.sendRequest<any, string, string, vscode.CancellationToken>(contentsRequest,
+    //            { uri: decodeURIComponent(uri.toString()) },
+    //            token);
+    //        return v || '';
+    //    }
+    //};
+    //context.subscriptions.push(vscode.workspace.registerTextDocumentContentProvider('jar', textDocumentContentProvider));
 
     // The title of this command is dictated by clojure-lsp and is executed when the user clicks the references code lens above a symbol
     context.subscriptions.push(vscode.commands.registerCommand('code-lens-references', async (_, line, character) => {